--- convert/util.go
@@ -1,99 +1,0 @@
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package convert
-
-import (
-	"fmt"
-	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
-	"strings"
-)
-
-var debugEnabled bool
-
-type Err struct {
-	Message string
-	File    string
-	Path    string
-	Func    string
-	Line    int
-}
-
-func (e *Err) Error() string {
-	return fmt.Sprintf("[%v:%v] %v", e.File, e.Line, e.Message)
-}
-
-func printTo(w io.Writer, i ...interface{}) {
-	s := fmt.Sprint(i...)
-	fmt.Fprintln(w, strings.TrimSuffix(s, "\n"))
-}
-
-func Warn(i ...interface{}) {
-	printTo(os.Stderr, i...)
-}
-
-func Info(i ...interface{}) {
-	printTo(os.Stderr, i...)
-}
-
-func Debug(i ...interface{}) {
-	if debugEnabled {
-		printTo(os.Stderr, i...)
-	}
-}
-
-func InitDebug() {
-	debugEnabled = true
-}
-
-func run(cmd *exec.Cmd) error {
-	/*if debugEnabled {
-		log.Printf("run: %v %v", cmd.Path, cmd.Args)
-	}*/
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return errorf(err.Error())
-	}
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		return errorf(err.Error())
-	}
-	go io.Copy(os.Stdout, stdout)
-	go io.Copy(os.Stderr, stderr)
-	return cmd.Run()
-}
-
-func errorf(format string, args ...interface{}) error {
-	msg := fmt.Sprintf(format, args...)
-	pc, filePath, lineNo, ok := runtime.Caller(1)
-	if !ok {
-		return &Err{
-			Message: msg,
-			File:    "unknown_file",
-			Path:    "unknown_path",
-			Func:    "unknown_func",
-			Line:    0,
-		}
-	}
-	return &Err{
-		Message: msg,
-		File:    filepath.Base(filePath),
-		Path:    filePath,
-		Func:    runtime.FuncForPC(pc).Name(),
-		Line:    lineNo,
-	}
-}
